Perform initial split and demo update
